# docs/openapi-specs.yaml
openapi: 3.0.3
info:
  title: Patient Management
  description: RESTful API for managing patients
  version: 1.0.0
  contact: {}
servers:
  - url: http://127.0.0.1:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/v1/visits:
    get:
      tags:
        - Visits
      summary: Get All Visits (Admin)
      description: |-
        **\--دریافت لیست تمام ویزیت ها--**

        امکانات:

        فیلتر کردن نتایج بر اساس مقدار

        محدود کردن فیلدهای خروجی

        محدود کردن تعداد نتایج

        مرتب سازی نتایج

        صفحه بندی نتایج
      operationId: getAllVisitsAdmin
      parameters:
        - name: fields
          in: query
          schema:
            type: string
            example: name,category,image
        - name: sort
          in: query
          schema:
            type: string
            example: category,-name
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: limit
          in: query
          schema:
            type: string
            example: '5'
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
    post:
      tags:
        - Visits
      summary: Create A Visit (Admin)
      description: |-
        **\--ایجاد یک ویزیت--**

        فیلدهای موجود:

        patient: آیدی بیمار (ضروری)

        doctor: آیدی دکتر (ضروری)

        closed: وضعیت فعال بودن ویزیت

        dateTime: تاریخ و زمان ویزیت (ضروری، در بازه زمانی یک ماه آینده)

        prescriptions: نسخه دکتر (شامل ساب فیلد)

        ‫فیلدهای موجود در prescriptions:

        drug: نام دارو

        count: تعداد دارو

        usage: نحوه مصرف دارو (حداکثر 200 کاراکتر)
      operationId: createAVisitAdmin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dateTime:
                  type: string
                  example: '2025-01-01T08:00:00.000+00:00'
                doctor:
                  type: string
                  example: doctor id
                patient:
                  type: string
                  example: patient id
            examples:
              Create A Visit (Admin):
                value:
                  dateTime: '2025-01-01T08:00:00.000+00:00'
                  doctor: doctor id
                  patient: patient id
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
  /api/v1/visits/patient:
    get:
      tags:
        - Visits
      summary: Get All My Visits (Patient)
      description: '**\--دریافت لیست ویزیت های شخصی بیمار--**'
      operationId: getAllMyVisitsPatient
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
    post:
      tags:
        - Visits
      summary: Book A Visit (Patient)
      description: >-
        **\--رزرو کردن یک ویزیت شخصی بیمار--**


        توجه 1: هر بیمار در هر روز تنها میتواند یک ویزیت با یک دکتر مشخص داشته
        باشد. رزرو کردن ویزیت با دکترهای دیگر مانعی ندارد.


        توجه 2: رزرو ویزیت در صورتی امکان پذیر است که اطلاعات حساب بیمار شامل
        ‫نام (name)، تاریخ تولد (birthday) و شماره ملی (idCard) تکمیل شده باشند.


        فیلدهای ضروری:


        doctor: آیدی دکتر


        dateTime: تاریخ و ساعت ویزیت (تنها مقادیر رند با ربع ساعت و مطابق با
        تنظیمات ویزیت دکتر قابل قبول است)
      operationId: bookAVisitPatient
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dateTime:
                  type: string
                  example: '2025-01-01T08:00:00.000+00:00'
                doctor:
                  type: string
                  example: doctor id
            examples:
              Book A Visit (Patient):
                value:
                  dateTime: '2025-01-01T08:00:00.000+00:00'
                  doctor: doctor id
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/visits/{_id}/patient:
    get:
      tags:
        - Visits
      summary: Get My Visit (Patient)
      description: '**\--دریافت یک ویزیت های شخصی بیمار بر اساس آیدی ویزیت--**'
      operationId: getMyVisitPatient
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
    delete:
      tags:
        - Visits
      summary: Delete My Visit (Patient)
      description: '**\--حذف کردن یک ویزیت شخصی بیمار بر اساس آیدی ویزیت--**'
      operationId: deleteMyVisitPatient
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
    parameters:
      - name: _id
        in: path
        required: true
        schema:
          type: string
          example: ''
        description: با مقدار آیدی جایگزین شود
...
  /api/v1/reviews:
    get:
      tags:
        - Reviews
      summary: Get All Reviews (Admin)
      description: |-
        **\--دریافت لیست تمام دیدگاه ها--**

        امکانات:

        فیلتر کردن نتایج بر اساس مقدار

        محدود کردن فیلدهای خروجی

        محدود کردن تعداد نتایج

        مرتب سازی نتایج

        صفحه بندی نتایج
      operationId: getAllReviewsAdmin
      parameters:
        - name: fields
          in: query
          schema:
            type: string
            example: name,category,image
        - name: sort
          in: query
          schema:
            type: string
            example: category,-name
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: limit
          in: query
          schema:
            type: string
            example: '5'
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
    post:
      tags:
        - Reviews
      summary: Create A Review (Admin)
      description: |-
        **\--ایجاد یک دیدگاه--**

        فیلدهای موجود:

        rating: امتیاز عددی (ضروری، عدد صحیح بین 1 تا 5)

        comment: دیدگاه متنی (ضروری، حداقل 10 کاراکتر، حداکثر 300 کاراکتر)

        patient: آیدی بیمار (ضروری)

        doctor: آیدی دکتر (ضروری)
      operationId: createAReviewAdmin
      requestBody:
        content:
          application/json:
            examples:
              Create A Review (Admin):
                value: ''
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...